+++ apps/rosen/app/(bridge)/WalletInfo.tsx
@@ -8,18 +8,28 @@ import {
   IconButton,
   Typography,
 } from '@rosen-bridge/ui-kit';
+import { Wallet } from '@rosen-ui/wallet-api';
+
+import { useWallet } from '@/_hooks';
 
 export interface WalletInfoProps {
   icon?: FunctionComponent;
   label?: string;
+  wallet?: Wallet;
   onClick: () => void;
 }
 
-export const WalletInfo = ({ icon: Icon, label, onClick }: WalletInfoProps) => {
+export const WalletInfo = ({
+  icon: Icon,
+  label,
+  wallet,
+  onClick,
+}: WalletInfoProps) => {
   const selected = Icon && label;
+  const { setSelected } = useWallet();
+
   return (
     <Grid container alignItems="center" justifyContent="space-between">
-      {/* <Typography lineHeight={2.5}>Wallet</Typography> */}
       {selected ? (
         <IconButton
           sx={{
@@ -50,10 +60,16 @@ export const WalletInfo = ({ icon: Icon, label, onClick }: WalletInfoProps) => {
         <Button
           sx={{
             color: (theme) => theme.palette.secondary.dark,
-            borderRadius: 2,
+            borderRadius: 1,
             margin: 0,
           }}
           startIcon={<Disconnect />}
+          onClick={() => {
+            if (wallet) {
+              wallet.disconnect();
+              setSelected?.(undefined);
+            }
+          }}
         >
           Disconnect
         </Button>