+++ apps/rosen/app/_backend/events/services.ts
@@ -1,3 +1,4 @@
+import { TokenMap } from '@rosen-bridge/tokens';
 import { Network } from '@rosen-ui/types';
 
 import { getTokenMap } from '@/_tokenMap/getServerTokenMap';
@@ -5,14 +6,17 @@ import { getTokenMap } from '@/_tokenMap/getServerTokenMap';
 import { UNSUPPORTED_TOKEN_NAME } from '../constants';
 import { getEvents } from './repository';
 
-const tokenMap = getTokenMap();
-
 /**
  * get full token data associated with a tokenId and a chain
+ * @param tokenMap
  * @param tokenId
  * @param chain
  */
-const getFullTokenData = (tokenId: string, chain: Network) => {
+const getFullTokenData = (
+  tokenMap: TokenMap,
+  tokenId: string,
+  chain: Network,
+) => {
   try {
     const token = tokenMap.search(chain, {
       [tokenMap.getIdKey(chain)]: tokenId,
@@ -43,13 +47,15 @@ export const getEventsWithFullTokenData = async (
   offset: number,
   limit: number,
 ) => {
+  const tokenMap = await getTokenMap();
+
   const events = await getEvents(offset, limit);
 
   return {
     total: events.total,
     items: events.items.map(({ sourceChainTokenId, ...item }) => ({
       ...item,
-      lockToken: getFullTokenData(sourceChainTokenId, item.fromChain),
+      lockToken: getFullTokenData(tokenMap, sourceChainTokenId, item.fromChain),
     })),
   };
 };