+++ networks/ergo/src/client.ts
@@ -4,17 +4,17 @@ import {
   NetworkConfig,
   NetworkMaxTransferParams,
 } from '@rosen-network/base';
-import { NETWORKS, NETWORK_LABELS } from '@rosen-ui/constants';
+import { NETWORKS } from '@rosen-ui/constants';
 import { RosenAmountValue } from '@rosen-ui/types';
 
 export class ErgoNetwork implements Network {
-  public label = NETWORK_LABELS.ERGO;
+  public label = NETWORKS.ergo.label;
 
   public lockAddress: string;
 
   public logo = ErgoIcon;
 
-  public name = NETWORKS.ERGO;
+  public name = NETWORKS.ergo.key;
 
   public nextHeightInterval: number;